passed process->pid to long int and removed get_next_pid's loop

+++ includes/corewar/process.h
@@ -6,7 +6,7 @@
 /*   By: abaurens <abaurens@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/21 12:57:27 by abaurens          #+#    #+#             */
-/*   Updated: 2019/11/09 03:56:41 by abaurens         ###   ########.fr       */
+/*   Updated: 2019/11/12 16:00:24 by abaurens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ struct			s_proc
 {
 	int			op;
 	uint16_t	pc;
-	uint32_t	pid;
+	long int	pid;
 	t_byte		carry;
 	uint32_t	lives;
 	uint32_t	last_live;
